# openapi.yaml
openapi: "3.0.3"
info:
  title: NiteLife API
  description: "This API provides an interface for managing events, ticket purchases, users and authentication."
  version: '1.0'
paths:
  /accounts:
    get:
      tags:
        - Accounts
      summary: "Get accounts"
      description: "retrieve all accounts from the database as per the filters and sorting criterias provided. Like all other paths that return an array, an **empty array** represents that a successful request that had no query results."
      parameters:
        - $ref: '#/components/parameters/filter'
        - $ref: '#/components/parameters/sort'
      responses:
        200:
          description: Successful Response.
          content:
            application/json:
              schema:
                title: Accounts
                type: array
                items:
                  $ref: '#/components/schemas/account'
        400:
          description: Request Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
    post:
      tags:
        - Accounts
      summary: "Create a new account"
      description: "Creates a new account with the given inputs"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "dima333"
                password:
                  type: string
                  example: "12345678"
                display_name:
                  type: string
                  example: "olgaLibber"
                full_name:
                  type: string
                  example: "Perry Platypus"
                gender:
                  type: string
                  enum: [ m, f, o ]
                  description: "The gender of the user. m = male, f = female, o = other"
                is_host:
                  type: boolean
      responses:
        201:
          description: "Successful Response"
        400:
          description: "Bad Request. Account was not created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
    account:
      description: "Represents a NiteLife user."
      type: object
      properties:
        id:
          type: integer
          example: 4444
        username:
          type: string
          example: "dima333"
        display_name:
          type: string
          example: "olgaLibber"
        full_name:
          type: string
          example: "Perry Platypus"
        gender:
          type: string
          enum: [ m, f, o ]
          description: "The gender of the user. m = male, f = female, o = other"
        is_host:
          type: boolean
...
    error:
      type: object
      description: "Response when request fails. The error field is to be used as a template message to display to users, while the messsage field is for debugging."
      properties:
        error:
          type: string
          example: "Bad request. Please try again"
        message:
          type: string
          example: "Input was invalid."
  parameters:
    filter:
      name: filter
      in: query
      description: "Set the filters to control what gets returned. See schema of object returned in successful responses. All fields of the returned object can be filtered as per the field's data type. **Default**: \"\"(no filters)"
      required: false
      schema:
        type: string
        example: "field1=value1&field2=value2"
    sort:
      name: sort
      in: query
      description: "Sorts the return values by the given field(s) and order. You can sort by multiple fields. Sort priority goes from left to right. +{field} sorts the results by the field in ascending order, -{field} is its descending equivalent.**Default**: \"\"(no sorting)"
      required: false
      schema:
        type: string
        example: "+field1&-field2"